# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "灵嘚中间件（LingDei-Middleware）是一个基于 Fiber 的 RESTful API 服务，用于提供灵嘚（LingDei）的后端服务。\n注意，有 🦸 标识的接口需要管理员权限才能访问。",
        "title": "灵嘚中间件",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
...
        "/comment/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加Comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "添加Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "评论内容",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/comment/count": {
            "get": {
                "description": "获取Comment数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "获取Comment数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommentCountResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommentCountResp"
                        }
                    }
                }
            }
        },
        "/comment/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除Comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "删除Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/comment/get": {
            "get": {
                "description": "获取Comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "获取Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommentResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/comment/list": {
            "get": {
                "description": "获取Comment列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "获取Comment列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommentListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommentListResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Comment": {
            "type": "object",
            "required": [
                "content",
                "timestamp",
                "user_uuid",
                "uuid",
                "video_uuid"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "profile": {
                    "$ref": "#/definitions/model.Profile"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video_uuid": {
                    "type": "string"
                }
            }
        },
        "model.CommentCountResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "model.CommentListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "comment_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Comment"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.CommentResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "comment": {
                    "$ref": "#/definitions/model.Comment"
                }
            }
        },
...
        "model.OperationResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Profile": {
            "type": "object",
            "required": [
                "id",
                "nickname"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}